// frontend/src/api/generated/model/activityItem.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * HuaXia Tourism RAG
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityAlternative } from './activityAlternative';
import type { ActivityItemCategory } from './activityItemCategory';

export interface ActivityItem {
  start_time?: string | null;
  end_time?: string | null;
  name: string;
  category?: ActivityItemCategory;
  description: string;
  location?: string | null;
  estimated_cost?: number | null;
  duration_hours?: number | null;
  booking_url?: string | null;
  opening_hours?: string | null;
  rating?: number | null;
  /** @maxItems 8 */
  citations?: number[];
  /** @maxItems 4 */
  alternatives?: ActivityAlternative[];
}